g_file_get_contents() takes a gsize not int (fixes bus error on 64-bit
2002-04-25  jacob berkman  <jacob@ximian.com>

	* src/session.c (load_state): g_file_get_contents() takes a gsize
	not int (fixes bus error on 64-bit platforms)

+++ src/session.c
@@ -976,7 +976,7 @@ load_state (const char *previous_save_file)
   GError *error;
   ParseData parse_data;
   char *text;
-  int length;
+  gsize length;
   char *session_file;
 
   session_file = g_strconcat (g_get_home_dir (),
